+++ packages/server/src/sdk/app/permissions/index.ts
@@ -1,10 +1,15 @@
+import { context, roles } from "@budibase/backend-core"
+import { features } from "@budibase/pro"
 import {
   DocumentType,
   PermissionLevel,
   VirtualDocumentType,
 } from "@budibase/types"
-import { isViewID } from "../../../db/utils"
-import { features } from "@budibase/pro"
+import { getRoleParams, isViewID } from "../../../db/utils"
+import {
+  CURRENTLY_SUPPORTED_LEVELS,
+  getBasePermissions,
+} from "../../../utilities/security"
 
 type ResourceActionAllowedResult =
   | { allowed: true }
@@ -34,4 +39,33 @@ export async function resourceActionAllowed({
     level,
     resourceType: VirtualDocumentType.VIEW,
   }
+}
+
+export async function getResourcePerms(resourceId: string) {
+  const db = context.getAppDB()
+  const body = await db.allDocs(
+    getRoleParams(null, {
+      include_docs: true,
+    })
+  )
+  const rolesList = body.rows.map(row => row.doc)
+  let permissions: Record<string, string> = {}
+  for (let level of CURRENTLY_SUPPORTED_LEVELS) {
+    // update the various roleIds in the resource permissions
+    for (let role of rolesList) {
+      const rolePerms = roles.checkForRoleResourceArray(
+        role.permissions,
+        resourceId
+      )
+      if (
+        rolePerms &&
+        rolePerms[resourceId] &&
+        rolePerms[resourceId].indexOf(level) !== -1
+      ) {
+        permissions[level] = roles.getExternalRoleID(role._id, role.version)!
+      }
+    }
+  }
+
+  return Object.assign(getBasePermissions(resourceId), permissions)
 }
